rename routes directory to features directory again
I think this keeps things clearer until we actually\/if adopt remix
