+++ client/app/features/ErrorPage/index.ts
@@ -1,0 +1,1 @@
+export { ErrorPage as Component } from '~/features/ErrorPage/ErrorPage';
