+++ client/app/features/ManifestDetails/ManifestDetails.tsx
