+++ client/app/features/ManifestDetails/index.ts
