+++ client/app/features/SiteList/index.ts
@@ -1,4 +1,4 @@
-import { SiteList } from '~/routes/SiteList/SiteList';
+import { SiteList } from '~/features/SiteList/SiteList';
 
 export { SiteList as Component };
 export default SiteList;
