+++ client/app/features/register/RegisterHero.tsx
