+++ client/app/features/register/index.ts
@@ -1,4 +1,4 @@
-import { RegisterHero } from '~/routes/register/RegisterHero';
+import { RegisterHero } from '~/features/register/RegisterHero';
 
 export { RegisterHero as Component };
 export default RegisterHero;
