--- client/app/routes/ErrorPage/index.ts
@@ -1,1 +1,0 @@
-export { ErrorPage as Component } from '~/routes/ErrorPage/ErrorPage';
